+++ package.json
@@ -31,7 +31,9 @@
     "@geolytix/xyz-app": "file:apps/xyz",
     "@geolytix/xyz-auth-app": "file:apps/auth",
     "@geolytix/xyz-saml-app": "file:apps/saml",
-    "express": "5.2.0"
+    "cookie-parser": "^1.4.7",
+    "express": "5.2.0",
+    "express-rate-limit": "^7.5.1"
   },
   "devDependencies": {
     "@biomejs/biome": "2.2.6",
